Added min_price and max_conf.
These will both give up if there either aren't enough workers or
allocations to service the entire batch. Perhaps it would be preferable
to allocate as many as possible. This means that the batch allocation UI
would have to be changed to accomodate half-allocated batches, which is
probably more work than it is worth.